Fix some Functions
- Fix ShowEvents
- Add a control in the DeleteEvent/Film/User/Prenotation functions()
- Fix the showUsersTable()
- Add CheckStringFK()
- Add the FK contol in DeleteEvent/Film/User/Prenotation functions()

+++ WCFClient/WCFClient/Pages/AdminMenu/DeleteEventPage.cs
@@ -21,29 +21,34 @@ namespace WCFClient.Pages
             Console.OutputEncoding = System.Text.Encoding.UTF8;
 
             try {
-            /*
-             * Show the Admin the Events
-             */
-                Output.WriteLine("EVENTS LIST: ");
-                TablePrinter.Event(SessionManager.GetServiceClient().GetEventsList());
-
-                /* 
-                 * Delete Event Form
-                 * 
-                 * Every Primary Key input must be valid.
-                 * When a Event is deleted, the Prenotations and the Reservations
-                 * linked to it are deleted too.
-                 */
-                Output.WriteLine("\n------ DELETE EVENT ------- ");
-                string event_code = Input.ReadString("Insert the Code of the Event to delete: ");
-                int eventCode = Controls.CheckInt(event_code);
-
                 /*
-                 * Send data to Database
+                 * Show the Admin the Events
                  */
-                if (SessionManager.GetServiceClient().DeleteEvent(eventCode))
-                    Output.WriteLine("\nEVENT CANCELLATION SUCCESS!\n");
-                else Output.WriteLine("\nEVENT CANCELLATION FAILED! Retry!\n");
+                if (SessionManager.GetServiceClient().GetEventsList().Count != 0)
+                {
+                    Output.WriteLine("EVENTS LIST: ");
+                    TablePrinter.Event(SessionManager.GetServiceClient().GetEventsList());
+
+                    /* 
+                     * Delete Event Form
+                     * 
+                     * Every Primary Key input must be valid.
+                     * When a Event is deleted, the Prenotations and the Reservations
+                     * linked to it are deleted too.
+                     */
+                    Output.WriteLine("\n------ DELETE EVENT ------- ");
+                    string event_code = Input.ReadString("Insert the Code of the Event to delete: ");
+                    int eventCode = Controls.CheckInt(event_code);
+                    eventCode = Controls.CheckIntForeignKey(eventCode.ToString(), "Evento");
+
+                    /*
+                     * Send data to Database
+                     */
+                    if (SessionManager.GetServiceClient().DeleteEvent(eventCode))
+                        Output.WriteLine("\nEVENT CANCELLATION SUCCESS!\n");
+                    else Output.WriteLine("\nEVENT CANCELLATION FAILED! Retry!\n");
+
+                } else Console.WriteLine("There are no Events in the DataBase!");
             }
             catch (Exception ex)
             {
