+++ WCFClient/WCFClient/Pages/AdminMenu/DeletePrenotationPage.cs
@@ -16,29 +16,35 @@ namespace WCFClient.Pages
 
             try
             {
-            /*
-             * Show the Admin the Prenotations
-             */
-                Output.WriteLine("PRENOTATIONS LIST: ");
-                TablePrinter.Prenotation(SessionManager.GetServiceClient().GetPrenotationsList());
+                if (SessionManager.GetServiceClient().GetPrenotationsList().Count != 0)
+                {
+                    /*
+                  * Show the Admin the Prenotations
+                  */
+                    Output.WriteLine("PRENOTATIONS LIST: ");
+                    TablePrinter.Prenotation(SessionManager.GetServiceClient().GetPrenotationsList());
 
-                /* 
-                 * Delete Film Form
-                 * 
-                 * Every Primary Key input must be valid.
-                 * When a Prenotation is deleted the Reservations 
-                 * linked to it are deleted too.
-                 */
-                Output.WriteLine("\n------ DELETE PRENOTATION ------- ");
-                string prenotation_code = Input.ReadString("Insert the Code of the Prenotation to delete: ");
-                int prenotationCode = Controls.CheckInt(prenotation_code);
+                    /* 
+                     * Delete Film Form
+                     * 
+                     * Every Primary Key input must be valid.
+                     * When a Prenotation is deleted the Reservations 
+                     * linked to it are deleted too.
+                     */
+                    Output.WriteLine("\n------ DELETE PRENOTATION ------- ");
+                    string prenotation_code = Input.ReadString("Insert the Code of the Prenotation to delete: ");
+                    int prenotationCode = Controls.CheckInt(prenotation_code);
+                    prenotationCode = Controls.CheckIntForeignKey(prenotationCode.ToString(), "Prenotazione");
 
-                /*
-                 * Send data to Database
-                 */
-                if (SessionManager.GetServiceClient().DeletePrenotation(prenotationCode))
-                    Output.WriteLine("\nPRENOTATION CANCELLATION SUCCESS!\n");
-                else Output.WriteLine("\nPRENOTATION CANCELLATION FAILED! Retry!\n");
+                    /*
+                     * Send data to Database
+                     */
+                    if (SessionManager.GetServiceClient().DeletePrenotation(prenotationCode))
+                        Output.WriteLine("\nPRENOTATION CANCELLATION SUCCESS!\n");
+                    else Output.WriteLine("\nPRENOTATION CANCELLATION FAILED! Retry!\n");
+
+                }
+                else Console.WriteLine("There are no Prenotations in the DataBase!");
             }
             catch (Exception ex)
             {
